Shamsi calendar, per-driver accounting report, and modern fast order entry

+++ app/src/main/java/moji/deliverytracker/OrderSummary.kt
@@ -1,6 +1,6 @@
 package moji.deliverytracker
 
 data class OrderSummary(
-    val total: Int,
+    val total: Long,
     val count: Int
 )
